Add test for classification
The test requires the result of taxizedb::classification is exactly the
same as taxize::classification.

To make this test pass, I had to remove the root node from the taxizedb
result. I think keeping the root node is actually the better choice,
since it is the common ancestor of cellular organisms and viruses.