Preserve input names to `classification`
Previously the names of the lineage list returned by `classification`
were always taxon IDs, even if the input were species names. Now the
names are exactly equal to the inputs (even if they are misspelled). For
example:

```
classification(c('pig', 'arabidopsis_thaliana'))
```

The implementation is a bit hacky, I'll clean it up soon.

I've also reordered the test inputs to catch any nonsense that crops up
if the names are names are sorted alphabetically but the list isn't.