Unnest hierarchy table
Previously the hierarchy table had the format:

```
tax_id   level   hierachy_string
9606     31      1-131567-....-9606
```

The hierarchy was stored in a string, which had to be parsed to get the
order. It would be a lot cleaner, and more in line with common database
practice, to unnest these values, e.g.

```
tax_id ancestor level
9606   1        31
9606   131567   30
9606   2759     29
...
9606   9605     1
9606   9606     0
```

In this format, it becomes trivial to get the full lineage of a taxa:

``` SQL
SELECT *
    FROM hierarchy
    WHERE tax_id IN (3702, 9606)
    ORDER BY tax_id, level DESC
;
```

Or to get all descendents of a set of taxa

``` SQL
SELECT tax_id, ancestor
    FROM hierarchy
    WHERE ancestor IN (3701, 9605)
;
```

The downside is that this breaks consistency with COL.

This commit breaks some existing functions in taxizedb; a problem I will
fix soon.